commented out the button on the playlistdisplay and moved track,tracklist to playlistdisplay folder
